+++ packages/astro/src/core/errors/index.ts
@@ -11,4 +11,4 @@ export {
 export type { ErrorLocation, ErrorWithMetadata } from './errors.js';
 export { codeFrame } from './printer.js';
 export { createSafeError, positionAt } from './utils.js';
-export { errorMap } from './error-map.js';
+export { errorMap } from './zod-error-map.js';
